Check if notification channels exist before deleting them (#990)

+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/PostgresNotificationSettingsAdapter.java
@@ -51,7 +51,9 @@ public class PostgresNotificationSettingsAdapter implements NotificationSettings
                         .map(channel -> new NotificationSettingsChannelEntity(userId.value(), entry.getKey(), channel)))
                 .toList();
 
-        notificationSettingsChannelRepository.deleteAllByUserId(userId.value());
+        if (notificationSettingsChannelRepository.existsByUserId(userId.value())) {
+            notificationSettingsChannelRepository.deleteAllByUserId(userId.value());
+        }
         notificationSettingsChannelRepository.saveAll(entities);
     }
 }
